/* fixed a huge bug with the "RemoveAttribute" method. First it needs to check if the attribute exists, before attempting to remove it. */

+++ server/mmud/characters/Person.java
@@ -895,7 +895,10 @@ public class Person implements Executable, AttributeContainer
 	{
 		Attribute attrib = getAttribute(aName);
 		theAttributes.remove(aName);
-		AttributeDb.removeAttribute(attrib, this);
+		if (attrib != null)
+		{
+			AttributeDb.removeAttribute(attrib, this);
+		}
 	}
 
 	/**
